BLADE-483 Copy manifest in tests and use it to read version
