+++ cli/src/main/java/com/liferay/blade/cli/BladeCLI.java
@@ -25,8 +25,9 @@ import com.beust.jcommander.Parameters;
 import com.liferay.blade.cli.command.BaseArgs;
 import com.liferay.blade.cli.command.BaseCommand;
 import com.liferay.blade.cli.command.BladeProfile;
+import com.liferay.blade.cli.command.UpdateCommand;
+import com.liferay.blade.cli.command.VersionCommand;
 import com.liferay.blade.cli.command.validator.ParameterPossibleValues;
-import com.liferay.blade.cli.util.BladeUtil;
 import com.liferay.blade.cli.util.CombinedClassLoader;
 import com.liferay.blade.cli.util.Prompter;
 
@@ -258,7 +259,7 @@ public class BladeCLI {
 			try {
 				_writeLastUpdateCheck();
 
-				BladeUtil.printUpdateIfAvailable(this);
+				_printUpdateIfAvailable();
 			}
 			catch (IOException ioe) {
 				error(ioe);
@@ -487,6 +488,10 @@ public class BladeCLI {
 					error(_jCommander.getParsedCommand() + ": " + pe.getMessage());
 				}
 			}
+		}
+		catch (Throwable e) {
+			error(e);
+			e.printStackTrace();
 		}
 		finally {
 			if (_extensionsClassLoaderSupplier != null) {
@@ -922,6 +927,49 @@ public class BladeCLI {
 				}
 			}
 		}
+	}
+
+	private boolean _printUpdateIfAvailable() throws IOException {
+		boolean available;
+
+		String bladeCLIVersion = VersionCommand.getBladeCLIVersion();
+
+		boolean fromSnapshots = false;
+
+		if (bladeCLIVersion == null) {
+			throw new IOException("Could not determine blade version");
+		}
+
+		fromSnapshots = bladeCLIVersion.contains("SNAPSHOT");
+
+		String updateVersion = "";
+
+		try {
+			updateVersion = UpdateCommand.getUpdateVersion(fromSnapshots);
+
+			available = UpdateCommand.shouldUpdate(bladeCLIVersion, updateVersion);
+
+			if (available) {
+				out(System.lineSeparator() + "blade version " + bladeCLIVersion + System.lineSeparator());
+				out(
+					"Run \'blade update" + (fromSnapshots ? " --snapshots" : "") + "\' to update to " +
+						(fromSnapshots ? "the latest snapshot " : " ") + "version " + updateVersion +
+							System.lineSeparator());
+			}
+			else {
+				if (fromSnapshots && !UpdateCommand.equal(bladeCLIVersion, updateVersion)) {
+					out(
+						String.format(
+							"blade version %s is newer than latest snapshot %s; skipping update.\n", bladeCLIVersion,
+							updateVersion));
+				}
+			}
+		}
+		catch (IOException ioe) {
+			available = false;
+		}
+
+		return available;
 	}
 
 	private String _promptForMissingParameter(Object commandArgs, Optional<String> missingParameterOptional) {
