+++ cli/src/main/java/com/liferay/blade/cli/command/UpdateArgs.java
@@ -47,7 +47,9 @@ public class UpdateArgs extends BaseArgs {
 		_url = url;
 	}
 
-	@Parameter(description = "Do not update, only check if an update is necessary.", hidden = true, names = {"-c", "--check"})
+	@Parameter(
+		description = "Do not update, only check if an update is necessary.", hidden = true, names = {"-c", "--check"}
+	)
 	private boolean _checkOnly;
 
 	@Parameter(description = "Switch to use the snapshot repository.", names = {"-s", "--snapshots"})
