+++ cli/src/main/java/com/liferay/blade/cli/util/BladeUtil.java
@@ -18,8 +18,6 @@ package com.liferay.blade.cli.util;
 
 import com.liferay.blade.cli.BladeCLI;
 import com.liferay.blade.cli.command.SamplesCommand;
-import com.liferay.blade.cli.command.UpdateCommand;
-import com.liferay.blade.cli.command.VersionCommand;
 import com.liferay.project.templates.ProjectTemplates;
 import com.liferay.project.templates.extensions.util.ProjectTemplatesUtil;
 
@@ -35,19 +33,11 @@ import java.io.PrintStream;
 import java.net.HttpURLConnection;
 import java.net.InetSocketAddress;
 import java.net.Socket;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.net.URL;
 
-import java.nio.MappedByteBuffer;
-import java.nio.channels.FileChannel;
 import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-
-import java.security.CodeSource;
-import java.security.MessageDigest;
-import java.security.ProtectionDomain;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -67,8 +57,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
-
-import javax.xml.bind.DatatypeConverter;
 
 /**
  * @author Gregory Amerson
@@ -219,29 +207,6 @@ public class BladeUtil {
 		}
 	}
 
-	public static String getMD5(Path path) {
-		try (FileChannel fileChannel = FileChannel.open(path)) {
-			long fileChannelSize = fileChannel.size();
-
-			MappedByteBuffer buffer = fileChannel.map(FileChannel.MapMode.READ_ONLY, 0, fileChannelSize);
-
-			MessageDigest messageDigest = MessageDigest.getInstance("MD5");
-
-			messageDigest.update(buffer);
-
-			byte[] digest = messageDigest.digest();
-
-			String md5Sum = DatatypeConverter.printHexBinary(digest);
-
-			buffer.clear();
-
-			return md5Sum.toUpperCase();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
 	public static Properties getProperties(File file) {
 		try (InputStream inputStream = new FileInputStream(file)) {
 			Properties properties = new Properties();
@@ -252,25 +217,6 @@ public class BladeUtil {
 		}
 		catch (Exception e) {
 			return null;
-		}
-	}
-
-	public static Path getRunningJarFile() {
-		try {
-			ProtectionDomain pd = BladeCLI.class.getProtectionDomain();
-
-			CodeSource cs = pd.getCodeSource();
-
-			URL location = cs.getLocation();
-
-			URI locationUri = location.toURI();
-
-			File runningJarFile = new File(locationUri);
-
-			return runningJarFile.toPath();
-		}
-		catch (URISyntaxException urise) {
-			throw new RuntimeException(urise);
 		}
 	}
 
@@ -377,49 +323,6 @@ public class BladeUtil {
 		}
 	}
 
-	public static boolean printUpdateIfAvailable(BladeCLI blade) throws IOException {
-		boolean available;
-
-		String bladeCLIVersion = VersionCommand.getBladeCLIVersion();
-
-		boolean fromSnapshots = false;
-
-		if (bladeCLIVersion == null) {
-			throw new IOException("Could not determine blade version");
-		}
-
-		fromSnapshots = bladeCLIVersion.contains("SNAPSHOT");
-
-		String updateVersion = "";
-
-		try {
-			updateVersion = UpdateCommand.getUpdateVersion(fromSnapshots);
-
-			available = UpdateCommand.shouldUpdate(bladeCLIVersion, updateVersion);
-
-			if (available) {
-				blade.out(System.lineSeparator() + "blade version " + bladeCLIVersion + System.lineSeparator());
-				blade.out(
-					"Run \'blade update" + (fromSnapshots ? " --snapshots" : "") + "\' to update to " +
-						(fromSnapshots ? "the latest snapshot " : " ") + "version " + updateVersion +
-							System.lineSeparator());
-			}
-			else {
-				if (fromSnapshots && !UpdateCommand.equal(bladeCLIVersion, updateVersion)) {
-					blade.out(
-						String.format(
-							"blade version %s is newer than latest snapshot %s; skipping update.\n", bladeCLIVersion,
-							updateVersion));
-				}
-			}
-		}
-		catch (IOException ioe) {
-			available = false;
-		}
-
-		return available;
-	}
-
 	public static String read(File file) throws IOException {
 		return new String(Files.readAllBytes(file.toPath()));
 	}
@@ -445,26 +348,6 @@ public class BladeUtil {
 
 		thread.setDaemon(true);
 		thread.start();
-	}
-
-	public static String readTextFileFromURL(String urlString) {
-		try {
-			StringBuilder sb = new StringBuilder();
-			URL url = new URL(urlString);
-
-			try (Scanner scanner = new Scanner(url.openStream())) {
-				while (scanner.hasNextLine()) {
-					sb.append(scanner.nextLine() + System.lineSeparator());
-				}
-			}
-
-			String returnValue = sb.toString();
-
-			return returnValue.trim();
-		}
-		catch (IOException ioe) {
-			throw new RuntimeException(ioe);
-		}
 	}
 
 	public static boolean searchZip(Path path, Predicate<String> test) {
