+++ cli/src/test/java/com/liferay/blade/cli/BladeCliTest.java
@@ -16,31 +16,42 @@
 
 package com.liferay.blade.cli;
 
-import com.liferay.blade.cli.util.BladeUtil;
+import java.io.File;
 
 import org.junit.Assert;
-import org.junit.Ignore;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 /**
+ * @author Christopher Bryan Boyd
  * @author Vernon Singleton
  */
 public class BladeCliTest {
 
-	@Ignore
+	@Before
+	public void setUp() throws Exception {
+		_rootDir = temporaryFolder.getRoot();
+
+		_extensionsDir = temporaryFolder.newFolder(".blade", "extensions");
+	}
+
 	@Test
 	public void testBladePrintUpdateIfAvailable() throws Exception {
-		BladeTest.BladeTestBuilder bladeTestBuilder = BladeTest.builder();
+		BladeTestResults results = TestUtil.runBlade(_rootDir, _extensionsDir, "update", "--check");
 
-		BladeTest bladeTest = bladeTestBuilder.build();
+		String output = results.getOutput();
 
-		Assert.assertTrue(
-			"New update should be available from snapshots.", BladeUtil.printUpdateIfAvailable(bladeTest));
+		boolean containsFalse = output.equals("false");
+
+		Assert.assertTrue("Current jar should be the latest version", containsFalse);
 	}
 
 	@Rule
 	public TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+	private File _extensionsDir = null;
+	private File _rootDir = null;
 
 }
